Add files via upload
Consider the IMDB example from Chapter 6. Re-run the example modifying the following:
The purpose of this assignment was to modify the IMDB dataset from the previous assignment (Assignment 1). We were told to make a cutoff reviews after 150 words, and restrict the training samples to only 100 while holding the Validate at 10,000 samples. And we were instructed to consider only the top 10,000 words. The main task was to consider both an embedding layer and a pre-trained word embedding. Finally, we were to determine which of the approach did better.  Next, we try changing the number of training samples to determine at what point the embedding layer gives better performance.